CiEEGData. Response2XLS, ResponseTriggerTime - added new parameter to XLS table - the length of significance.
CiEEGData. ResponseTriggerTime - computes also the total length of significance - relative to baseline and also between the categories. Is returned in the len vector.
		- correction of WpAllCh for differences relative to baseline.  The WpK was zero, so redundant.
	- Response2XLS takes the len values and saves them to new columns in XLS
cIntegrate - returns also the total area under curve (in the time range of significance. Althoutgh it is not used now
CHHeader . PacientTag - forcecell parameter to return the cell array even if there is only one channel. Is used in Response2XLS